new methods

+++ src/services/Photo/index.js
@@ -2,6 +2,9 @@ const {
   ServiceError
 } = require("../../errors");
 const wrapServiceAction = require("../_core/wrapServiceAction");
+const checkAuthorization = require("../_core/checkAuthorization");
+
+const pick = require("lodash/pick");
 
 const utils = require("../../utils");
 
@@ -19,25 +22,42 @@ const { string, any } = require("../../validation");
 /*
 * Service Actions
 * */
-module.exports.createPhoto = wrapServiceAction({
+module.exports.createPhotos = wrapServiceAction({
   params: {
     $$strict: "remove",
+    accountId: { ...any },
     ownerId: { ...any },
     ownerType: {
       type: "enum",
       values: ["location", "project", "event"]
     },
-    filename: { ...string },
-    description: { ...string, optional: true }
+    photos: {
+      type: "array",
+      min: 1,
+      items: {
+        type: "object",
+        props: {
+          filename: { ...string },
+          description: {
+            ...string,
+            optional: true
+          }
+        }
+      }
+    }
   },
   async handler (params) {
-    const photo = await models.Photo.create({
-      ...params
-    });
-    await models.PendingUpload.deleteOne({
-      filename: params.filename
-    });
-    return photo;
+    await checkAuthorization(params.accountId, params.ownerId, params.ownerType);
+    for (const photo of params.photos) {
+      await models.Photo.create({
+        ...pick(params, ["accountId", "ownerId", "ownerType"]),
+        ...photo
+      });
+      await models.PendingUpload.deleteOne({
+        filename: photo.filename
+      });
+    }
+    return true;
   }
 });
 
@@ -56,9 +76,11 @@ module.exports.getPhotos = wrapServiceAction({
     });
   }
 });
+
 module.exports.deletePhoto = wrapServiceAction({
   params: {
     $$strict: "remove",
+    accountId: { ...any },
     photoId: { ...any }
   },
   async handler (params) {
@@ -66,6 +88,7 @@ module.exports.deletePhoto = wrapServiceAction({
     if (!photo) {
       throw new ServiceError("photo not found");
     }
+    await checkAuthorization(params.accountId, photo.ownerId, photo.ownerType);
     await utils.deleteUploadedFile(photo.filename).catch(console.error);
     await photo.delete();
     return true;
